+++ src/apis/QueueLogController.php
@@ -10,7 +10,7 @@ namespace luya\admin\apis;
 class QueueLogController extends \luya\admin\ngrest\base\Api
 {
     /**
-     * @var string The path to the model which is the provider for the rules and fields.
+     * {@inheritDoc}
      */
     public $modelClass = 'luya\admin\models\QueueLog';
 }