+++ src/controllers/LoggerController.php
@@ -2,6 +2,7 @@
 
 namespace luya\admin\controllers;
 
+use luya\admin\Module;
 use luya\admin\ngrest\base\Controller;
 
 /**
@@ -12,5 +13,21 @@ use luya\admin\ngrest\base\Controller;
  */
 class LoggerController extends Controller
 {
+    /**
+     * {@inheritDoc}
+     */
     public $modelClass = 'luya\admin\models\Logger';
+
+    /**
+     * {@inheritDoc}
+     */
+    public $clearButton = true;
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getDescription()
+    {
+        return Module::t('logger_controller_description');
+    }
 }