+++ src/messages/nl/admin.php
@@ -172,7 +172,7 @@ return [
     'ngrest_crud_group_prompt' => 'Groeperen op veld',
     'ngrest_crud_filter_prompt' => 'Selecteer filter',
     'ngrest_crud_ajax_search_length' => 'De zoekopdracht moet een minimale lengte van 3 tekens hebben.',
-    'menu_system_logger' => 'Loggen',
+    'menu_system_logger' => 'Application Loggen',
     'layout_filemanager_detail_replace_file' => 'Bestand vervangen',
     'layout_rightbar_languagelabel' => 'Interfacetaal',
     'layout_rightbar_savebtn' => 'Opslaan',
@@ -490,4 +490,14 @@ return [
     'crop_quality_high' => 'High',
     'crop_quality_medium' => 'Medium',
     'crop_quality_low' => 'Low',
+
+// 3.2.0
+
+    'menu_system_model_event_log' => 'Model Event Log',
+    'ngrestlog_controller_description' => 'Model Event Log entries are captured when creating, updating or deleting records in CRUD views of the admin.',
+    'logger_controller_description' => 'Application Log entries are triggered by [Logger Model](https://luya.io/api/luya-admin-models-Logger).',
+    'config_controller_description' => 'Create configuration keys and assign a value using [Config Model](https://luya.io/api/luya-admin-models-Config).',
+    'queuelog_controller_description' => 'Lists all unprocessed queue jobs, read [Queue Guide](https://luya.io/guide/app-queue) for a proper setup.',
+    'ngrest_delete_all_button_label' => 'Delete all data',
+    'ngrest_delete_all_button_confirm_message' => 'Are you sure to delete all data? This can not be undone!',
 ];