+++ src/ngrest/base/Controller.php
@@ -4,6 +4,8 @@ namespace luya\admin\ngrest\base;
 
 use Yii;
 use Exception;
+use luya\admin\components\Auth;
+use luya\admin\Module;
 use yii\base\InvalidConfigException;
 use luya\admin\ngrest\NgRest;
 use luya\admin\ngrest\render\RenderCrud;
@@ -13,7 +15,8 @@ use yii\web\ForbiddenHttpException;
 /**
  * Base Controller for all NgRest Controllers.
  *
- * @property \luya\admin\ngrest\base\NgRestModel $model The model based from the modelClass instance
+ * @property NgRestModel $model The model based from the modelClass instance
+ * @property string $description A text descirption for the CRUD which is display below the CRUD title
  *
  * @author Basil Suter <basil@nadar.io>
  * @since 1.0.0
@@ -79,6 +82,14 @@ class Controller extends \luya\admin\base\Controller
     public $globalButtons = [];
 
     /**
+     * @var boolean Whether the CRUD dropdown shows a button with "Delete all data" which will then truncate the whole table content. Usefull for tables
+     * which contain only temporary informations like logs and errors. If the user does not have delete permission, the button is hidden. In order to
+     * make the clear buutton (truncates all the data) work, the {{luya\admin\ngrest\base\Api::$truncateAction}} must be enabled.
+     * @since 3.0.0
+     */
+    public $clearButton = false;
+
+    /**
      * @var string|array|\luya\admin\ngrest\render\RenderCrudInterface
      *
      * You can customize crud rendering using this property. To do so you may use standard `\luya\admin\ngrest\render\RenderCrud`, but customize
@@ -109,6 +120,30 @@ class Controller extends \luya\admin\base\Controller
      * @since 2.0.0
      */
     public $renderCrud = RenderCrud::class;
+
+    private $_description;
+
+    /**
+     * Setter method for description
+     *
+     * @param string $description
+     * @since 3.2.0
+     */
+    public function setDescription($description)
+    {
+        $this->_description = $description;
+    }
+
+    /**
+     * Getter method for description
+     *
+     * @return string
+     * @since 3.2.0
+     */
+    public function getDescription()
+    {
+        return $this->_description;
+    }
     
     /**
      * @inheritdoc
@@ -163,9 +198,19 @@ class Controller extends \luya\admin\base\Controller
         if ($userSortSettings && is_array($userSortSettings) && $config->getDefaultOrder() !== false) {
             $config->defaultOrder = [$userSortSettings['field'] => $userSortSettings['sort']];
         }
+
+        // check if delete permissions exists for the current user in this crud.
+        if ($this->clearButton && Yii::$app->auth->matchApi(Yii::$app->adminuser->getId(), $config->apiEndpoint, Auth::CAN_DELETE)) {
+            array_push($this->globalButtons, [
+                'icon' => 'delete',
+                'label' => Module::t('ngrest_delete_all_button_label'),
+                'ng-click' => "clearData()"
+            ]);
+        }
         
         // generate crud renderer
         $crud = Yii::createObject($this->renderCrud);
+        $crud->description = $this->description;
         $crud->setModel($this->model);
         $crud->setSettingButtonDefinitions($this->globalButtons);
         $crud->setIsInline($inline);