+++ src/resources/js/controllers.js
@@ -27,6 +27,16 @@
 
 		$scope.tabService = CrudTabService;
 
+
+		$scope.clearData = function() {
+			AdminToastService.confirm(i18n['ngrest_delete_all_button_confirm_message'], i18n['ngrest_delete_all_button_label'], function() {
+				var toast = this;
+				$http.get($scope.config.apiEndpoint + '/truncate').then(function() {
+					toast.close();
+					$scope.loadList();
+				});
+			});
+		};
 		/***** TABS AND SWITCHES *****/
 
 		/**